test(analytics): share the ClickHouse repository fakes, correct a stale test doc
Three review comments on the review commits themselves.

The two repository suites had byte-identical copies of the ORDER BY-applying
fake and its sort helper — the same duplication rule the production decoders
were held to, and worse here, because a parser that drifted in one copy would
quietly stop proving the tiebreak in that suite while still proving it in the
other. Extracted to a shared module and re-verified by mutation: deleting the
repository's ORDER BY still fails both tie tests through the shared fake.

The fold-equivalence suite still carried a "KNOWN FAILURE … left red on purpose"
block for the prompt_ids divergence, which this same PR fixes — the comment
described behaviour the code no longer has. Rewritten to say what the fixture
now guards, including that nine spans sits just past the OLD 256-char boundary
and deliberately does not reach the 4096-char one.

Also switched the event fixture's envelope() to named parameters.
