+++ langwatch/src/server/event-sourcing/pipelines/evaluation-processing/projections/__tests__/fixtures/evaluation-events.fixtures.ts
@@ -43,11 +43,15 @@ interface OutcomeOptions {
   costId?: string;
 }
 
-function envelope(
-  type: string,
-  options: CommonOptions,
-  fallbackEventId: string,
-) {
+function envelope({
+  type,
+  options,
+  fallbackEventId,
+}: {
+  type: string;
+  options: CommonOptions;
+  fallbackEventId: string;
+}) {
   return {
     type,
     id: options.eventId ?? fallbackEventId,
@@ -62,7 +66,7 @@ export function createEvaluationScheduledEvent(
   options: IdentityOptions = {},
 ): EvaluationScheduledEvent {
   return {
-    ...envelope("lw.evaluation.scheduled", options, "evt-1"),
+    ...envelope({ type: "lw.evaluation.scheduled", options, fallbackEventId: "evt-1" }),
     data: {
       evaluationId: options.evaluationId ?? DEFAULT_EVALUATION_ID,
       evaluatorId: options.evaluatorId ?? "monitor-x",
@@ -78,7 +82,7 @@ export function createEvaluationStartedEvent(
   options: IdentityOptions = {},
 ): EvaluationStartedEvent {
   return {
-    ...envelope("lw.evaluation.started", options, "evt-2"),
+    ...envelope({ type: "lw.evaluation.started", options, fallbackEventId: "evt-2" }),
     data: {
       evaluationId: options.evaluationId ?? DEFAULT_EVALUATION_ID,
       evaluatorId: options.evaluatorId ?? "monitor-x",
@@ -98,7 +102,7 @@ export function createEvaluationCompletedEvent(
   options: CommonOptions & OutcomeOptions = {},
 ): EvaluationCompletedEvent {
   return {
-    ...envelope("lw.evaluation.completed", options, "evt-3"),
+    ...envelope({ type: "lw.evaluation.completed", options, fallbackEventId: "evt-3" }),
     data: {
       evaluationId: options.evaluationId ?? DEFAULT_EVALUATION_ID,
       status: options.status ?? "processed",
@@ -117,7 +121,7 @@ export function createEvaluationReportedEvent(
   options: IdentityOptions & OutcomeOptions = {},
 ): EvaluationReportedEvent {
   return {
-    ...envelope("lw.evaluation.reported", options, "evt-r"),
+    ...envelope({ type: "lw.evaluation.reported", options, fallbackEventId: "evt-r" }),
     data: {
       evaluationId: options.evaluationId ?? DEFAULT_EVALUATION_ID,
       evaluatorId: options.evaluatorId ?? "monitor-y",
